rename dashboard to api and place auth on api rout
